Fix an issue when calculating the line and address diff
git-svn-id: https://smali.googlecode.com/svn/trunk@361 55b6fa8a-2a1e-11de-a435-ffa8d773f76a

+++ dexlib/src/main/java/org/jf/dexlib/Debug/DebugInstructionIterator.java
@@ -231,7 +231,7 @@ public class DebugInstructionIterator {
                 }
                 default:
                 {
-                    byte base = (byte)((debugOpcode & 0xFF) - 0x0A);
+                    int base = ((debugOpcode & 0xFF) - 0x0A);
                     address += base / 15;
                     line += (base % 15) - 4;
                     processDecodedDebugInstruction.ProcessLineEmit(address, line);
